perf(sqlite): trim snapshot child import graph
