+++ config/assertion-safety-baseline.txt
@@ -3061,7 +3061,6 @@ src/infra/sqlite-readonly-location.ts	3
 src/infra/sqlite-schema-contract.ts	9
 src/infra/sqlite-snapshot.ts	2
 src/infra/sqlite-strict.ts	7
-src/infra/sqlite-transaction.ts	2
 src/infra/sqlite-user-version.ts	1
 src/infra/sqlite-wal.ts	5
 src/infra/startup-migration-checkpoint.ts	1
