+++ src/cli/update-cli/update-command-result-diagnostics.test.ts
@@ -3,7 +3,7 @@ import { afterEach, expect, it } from "vitest";
 import { useAutoCleanupTempDirTracker } from "../../../test/helpers/temp-dir.js";
 import { formatErrorMessage } from "../../infra/errors.js";
 import { openNodeSqliteDatabase } from "../../infra/node-sqlite.js";
-import { isSqliteLockError } from "../../infra/sqlite-transaction.js";
+import { isSqliteLockError } from "../../infra/sqlite-error-diagnostics.js";
 import { formatUpdateFinalizationError } from "./update-command-result.js";
 
 const dirs = useAutoCleanupTempDirTracker(afterEach);
