+++ src/commands/doctor-lint.test.ts
@@ -37,8 +37,8 @@ vi.mock("../config/config.js", async (importOriginal) => {
     readConfigFileSnapshot: mocks.readConfigFileSnapshot,
   };
 });
-vi.mock("../infra/sqlite-readonly-location.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../infra/sqlite-readonly-location.js")>();
+vi.mock("../infra/sqlite-snapshot-source.js", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("../infra/sqlite-snapshot-source.js")>();
   mocks.actualPrepareSqliteReadOnlyLocationSync.mockImplementation(
     actual.prepareSqliteReadOnlyLocationSync,
   );
