+++ src/infra/sqlite-readonly-location.ts
@@ -13,12 +13,7 @@ import {
   createPrivateSqliteTempDirectorySync,
   resolvePrivateSqliteSnapshotStagingRoot,
 } from "./sqlite-private-directory.js";
-import { runSqliteReadOnlyWorker, runSqliteReadOnlyWorkerSync } from "./sqlite-readonly-worker.js";
 import { withSqliteSourceHandle, withSqliteSourceHandleAsync } from "./sqlite-source-handle.js";
-import {
-  hasStateDatabaseSourceExclusion,
-  prepareStateDatabaseMutationSnapshot,
-} from "./state-database-coordinator.js";
 
 const MAX_SNAPSHOT_ATTEMPTS = 10;
 const COPY_BUFFER_BYTES = 1024 * 1024;
@@ -26,7 +21,7 @@ const SQLITE_HEADER_BYTES = 20;
 const SQLITE_READONLY_RESULT_CODE = 8;
 const SQLITE_RESULT_CODE_MASK = 0xff;
 const SQLITE_JOURNAL_MAGIC = Buffer.from([0xd9, 0xd5, 0x05, 0xf9, 0x20, 0xa1, 0x63, 0xd7]);
-const SQLITE_SNAPSHOT_STAGING_PREFIX = `openclaw-sqlite-readonly-${process.pid}-`;
+export const SQLITE_SNAPSHOT_STAGING_PREFIX = `openclaw-sqlite-readonly-${process.pid}-`;
 const pendingTempDirectoryCleanup = new Set<string>();
 let cleanupExitHandlerInstalled = false;
 
@@ -44,7 +39,7 @@ type SourceSidecars = {
 
 type SourceJournalMode = "empty" | "rollback" | "unknown" | "wal";
 
-type PreparedSqliteReadOnlyLocation = {
+export type PreparedSqliteReadOnlyLocation = {
   cleanup: () => boolean;
   location: string;
 };
@@ -299,7 +294,7 @@ function rollbackJournalReferencesSuperJournal(journalPath: string): boolean {
   }
 }
 
-function removeTempDirectory(tempDir: string): boolean {
+export function removeTempDirectory(tempDir: string): boolean {
   try {
     fs.rmSync(tempDir, { force: true, maxRetries: 3, recursive: true, retryDelay: 20 });
     pendingTempDirectoryCleanup.delete(tempDir);
@@ -322,7 +317,7 @@ function removeTempDirectory(tempDir: string): boolean {
   }
 }
 
-function adoptPreparedLocation(
+export function adoptPreparedLocation(
   location: string,
   ownedRoot?: string,
   requireCleanup = false,
@@ -432,7 +427,7 @@ function createStableReadOnlyCopyInTempDirectory(
   }
 }
 
-async function createSqliteSnapshotStagingDirectory(
+export async function createSqliteSnapshotStagingDirectory(
   stagingRoot = resolvePrivateSqliteSnapshotStagingRoot(),
 ): Promise<string> {
   try {
@@ -616,82 +611,6 @@ function prepareReadOnlySourceSyncInProcess(
   throw new Error(`SQLite source did not stabilize for read-only inspection: ${canonicalPath}`, {
     cause: lastChange,
   });
-}
-
-export async function prepareSqliteReadOnlyLocation(
-  pathname: string,
-  options: { preserveSourceArtifacts?: boolean; signal?: AbortSignal } = {},
-): Promise<PreparedSqliteReadOnlyLocation> {
-  let stagingRoot: string | undefined;
-  try {
-    options.signal?.throwIfAborted();
-    const ownedSnapshot = prepareStateDatabaseMutationSnapshot(pathname);
-    if (ownedSnapshot) {
-      const prepared = await ownedSnapshot;
-      try {
-        options.signal?.throwIfAborted();
-        return prepared;
-      } catch (error) {
-        prepared.cleanup();
-        throw error;
-      }
-    }
-    if (hasStateDatabaseSourceExclusion(pathname)) {
-      const prepared = options.preserveSourceArtifacts
-        ? prepareSqliteReadOnlyLocationSyncInProcess(pathname)
-        : await prepareSqliteReadOnlyLocationInProcess(pathname);
-      try {
-        options.signal?.throwIfAborted();
-        return prepared;
-      } catch (error) {
-        prepared.cleanup();
-        throw error;
-      }
-    }
-    // A stopped worker may never publish its random snapshot path. Allocate its
-    // private parent first so cancellation can join the child and remove all copies.
-    stagingRoot = await createSqliteSnapshotStagingDirectory();
-    options.signal?.throwIfAborted();
-    const location = await runSqliteReadOnlyWorker(pathname, {
-      mode: options.preserveSourceArtifacts ? "sync" : "async",
-      signal: options.signal,
-      stagingRoot,
-    });
-    options.signal?.throwIfAborted();
-    // Cancellable maintenance must retain its fence on cleanup failure; ordinary
-    // read-only handles report false so their owner can retry close.
-    return adoptPreparedLocation(location, stagingRoot, options.signal !== undefined);
-  } catch (error) {
-    if (stagingRoot && !removeTempDirectory(stagingRoot)) {
-      throw new Error(`SQLite read-only worker snapshot cleanup failed: ${stagingRoot}`, {
-        cause: error,
-      });
-    }
-    options.signal?.throwIfAborted();
-    throw error;
-  }
-}
-
-export function prepareSqliteReadOnlyLocationSync(
-  pathname: string,
-): PreparedSqliteReadOnlyLocation {
-  if (hasStateDatabaseSourceExclusion(pathname)) {
-    return prepareSqliteReadOnlyLocationSyncInProcess(pathname);
-  }
-  const stagingRoot = createPrivateSqliteTempDirectorySync(
-    resolvePrivateSqliteSnapshotStagingRoot(),
-    SQLITE_SNAPSHOT_STAGING_PREFIX,
-  );
-  try {
-    return adoptPreparedLocation(runSqliteReadOnlyWorkerSync(pathname, stagingRoot), stagingRoot);
-  } catch (error) {
-    if (!removeTempDirectory(stagingRoot)) {
-      throw new Error(`SQLite read-only worker snapshot cleanup failed: ${stagingRoot}`, {
-        cause: error,
-      });
-    }
-    throw error;
-  }
 }
 
 export function prepareSqliteReadOnlyLocationInProcess(pathname: string, stagingRoot?: string) {
