+++ src/state/openclaw-state-db-readonly.test.ts
@@ -4,7 +4,7 @@ import path from "node:path";
 import { DatabaseSync } from "node:sqlite";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { hasNodeErrorCode } from "../infra/path-guards.js";
-import * as sqliteReadOnly from "../infra/sqlite-readonly-location.js";
+import * as sqliteReadOnly from "../infra/sqlite-snapshot-source.js";
 import { createDeferredCore } from "../shared/deferred.js";
 import { withOpenClawTestState } from "../test-utils/openclaw-test-state.js";
 import { withTempDir } from "../test-utils/temp-dir.js";
